add tests for methods in user service

+++ src/main/java/com/reservation_restaurants_service/dto/RestaurantDto.java
@@ -19,5 +19,5 @@ public class RestaurantDto {
     private String address;
     @NotBlank(message = "Field cannot be empty")
     private long phoneNumber;
-    private double rating;
+    private long rating;
 }
